python-files/requirements.txt
aiogram==3.15.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
aiohttp==3.10.10
alembic==1.14.0
thefuzz==0.22.1
Levenshtein==0.26.0